Changes to Creating Appointment and some fixes with writing information to database
Changed creating appointment. Now the user has the ability to pick a day from the first 90 available which have less then the max available vacinations per day.
Also fixed some bugs with writing to the databases and made it work with the new way of creating appointments.
Fixed confirm appointment button and function for doctors.
When user has missed his appointment, he is able to create a new one, even though his exists in the database.