Error detection for appointment and visual changes
When creating an appointment, if someone created an appointment before you clocked confirm, and the other user got the last available appointment at that day, an error appears.
When a logged in users goes to the home page, instead of their email, their full name displays on the page.
Changed buttons in navbar to have the font color white, to match the previously changed navbar background color.